tests/wpt/web-platform-tests/bluetooth/server/device-same-object.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return getDiscoveredHealthThermometerDevice()
    .then(({device, fake_peripheral}) => {
      return fake_peripheral.setNextGATTConnectionResponse({code: HCI_SUCCESS})
        .then(() => device.gatt.connect());
    })
    .then(gatt => {
      assert_equals(gatt.device, gatt.device);
    });
}, "[SameObject] test for BluetoothRemoteGATTServer's device.");
</script>
